+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## NEXT
 
-- Removed `**kwargs` from the `add_node`, `add_edge`, and `add_edge_by_pair` methods of
-  `cai_causal_graph.causal_graph.CausalGraph` and its subclasses.
+- Removed `**kwargs` from the `cai_causal_graph.causal_graph.CausalGraph.add_node`,
+  `cai_causal_graph.causal_graph.CausalGraph.add_edge`, and `cai_causal_graph.causal_graph.CausalGraph.add_edge_by_pair`
+  methods in `cai_causal_graph.causal_graph.CausalGraph` and its subclasses.
 
 ## 0.3.14